Upgrade TypeScript to v6 and remove ts-jest
