+++ babel.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  presets: [
+    ['@babel/env', { targets: { node: 'current' } }],
+    ['@babel/preset-typescript', { allowDeclareFields: true }],
+  ],
+}
