+++ jest.config.mjs
@@ -1,7 +1,3 @@
-import { createDefaultPreset } from 'ts-jest'
-
-const tsjPreset = createDefaultPreset()
-
 /** @type {import('jest').Config} */
 const config = {
   collectCoverageFrom: ['src/**/*.{j,t}s'],
@@ -9,9 +5,10 @@ const config = {
   coverageThreshold: { global: { lines: 95 } },
   setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
   testEnvironment: 'node',
-  testRegex: '(/(test|__tests__)/(?![_.]).*|(\\.|/)(test|spec))\\.[jt]s$',
+  testRegex:
+    '(/(test|__tests__)/(?![_.]).*|(\\.|/)(test|spec))(?<!\\.d)\\.[jt]s$',
   transform: {
-    ...tsjPreset.transform,
+    '^.+\\.[mc]?[tj]s$': 'babel-jest',
     '^.*\\.s[ac]ss$': '<rootDir>/test/_transformers/sass.js',
   },
   prettierPath: null,
