+++ package.json
@@ -29,7 +29,7 @@
   "types": "types/src/marp.d.ts",
   "files": [
     "lib/",
-    "types/",
+    "types/src",
     "browser.js",
     "browser.d.ts"
   ],
@@ -65,6 +65,9 @@
     "exec": "npx @marp-team/marp-cli@latest -y"
   },
   "devDependencies": {
+    "@babel/core": "^7.29.7",
+    "@babel/preset-env": "^7.29.7",
+    "@babel/preset-typescript": "^7.29.7",
     "@csstools/postcss-minify": "^2.0.5",
     "@rollup/plugin-alias": "^6.0.0",
     "@rollup/plugin-commonjs": "^29.0.3",
@@ -77,6 +80,7 @@
     "@twemoji/api": "^17.0.3",
     "@types/jest": "^30.0.0",
     "autoprefixer": "^10.5.2",
+    "babel-jest": "^30.4.1",
     "cheerio": "^1.2.0",
     "cssnano": "^8.0.2",
     "emoji-regex": "10.6.0",
@@ -108,9 +112,8 @@
     "self-closing-tags": "^1.0.1",
     "stylelint": "^17.14.0",
     "stylelint-config-standard-scss": "^17.0.0",
-    "ts-jest": "29.4.11",
     "tslib": "^2.8.1",
-    "typescript": "^5.9.3",
+    "typescript": "^6.0.3",
     "typescript-eslint": "^8.62.0"
   },
   "dependencies": {
